refactor : 자유게시판 댓글기능

+++ src/main/java/com/example/coffies_vol_02/commnet/domain/dto/request/CommentRequest.java
@@ -12,6 +12,6 @@ public record CommentRequest(
         @NotBlank
         String replyContents,
         @Schema(name = "replyPoint",description = "댓글 평점",type = "Integer")
-        Integer replyPoint) {
+        Double replyPoint) {
 
 }
